+++ packages/cspell-lib/src/lib/Settings/Controller/configLoader/configLoader.ts
@@ -41,6 +41,7 @@ import { pnpLoader } from '../pnpLoader.js';
 import { searchPlaces } from './configLocations.js';
 import { ConfigSearch } from './configSearch.js';
 import { configToRawSettings } from './configToRawSettings.js';
+import { StopSearchAt } from './defaultConfigLoader.js';
 import { defaultSettings } from './defaultSettings.js';
 import {
     normalizeCacheSettings,
@@ -83,6 +84,12 @@ interface ImportedConfigEntry {
     referencedSet: Set<string>;
 }
 
+export interface SearchForConfigFileOptions {
+    stopSearchAt?: StopSearchAt;
+}
+
+export interface SearchForConfigOptions extends SearchForConfigFileOptions, PnPSettingsOptional {}
+
 let defaultConfigLoader: ConfigLoaderInternal | undefined = undefined;
 
 interface CacheMergeConfigFileWithImports {
@@ -113,12 +120,12 @@ export interface IConfigLoader {
     /**
      * This is an alias for `searchForConfigFile` and `mergeConfigFileWithImports`.
      * @param searchFrom the directory / file URL to start searching from.
-     * @param pnpSettings - related to Using Yarn PNP.
+     * @param options - Optional settings including stop location and Yarn PnP configuration.
      * @returns the resulting settings
      */
     searchForConfig(
         searchFrom: URL | string | undefined,
-        pnpSettings?: PnPSettingsOptional,
+        options?: SearchForConfigOptions,
     ): Promise<CSpellSettingsI | undefined>;
 
     resolveConfigFileLocation(filenameOrURL: string | URL, relativeTo?: string | URL): Promise<URL | undefined>;
@@ -216,6 +223,7 @@ export class ConfigLoader implements IConfigLoader {
     protected globalSettings: CSpellSettingsI | undefined;
     protected cspellConfigFileReaderWriter: CSpellConfigFileReaderWriter;
     protected configSearch: ConfigSearch;
+    protected stopSearchAtCache: WeakMap<SearchForConfigFileOptions, URL[] | undefined> = new WeakMap();
 
     protected toDispose: { dispose: () => void }[] = [];
 
@@ -255,21 +263,19 @@ export class ConfigLoader implements IConfigLoader {
         });
     }
 
-    async searchForConfigFileLocation(searchFrom: URL | string | undefined): Promise<URL | undefined> {
-        const url = toFileURL(searchFrom || cwdURL(), cwdURL());
-        if (
-            typeof searchFrom === 'string' &&
-            !isUrlLike(searchFrom) &&
-            url.protocol === 'file:' && // check to see if it is a directory
-            (await isDirectory(this.fs, url))
-        ) {
-            return this.configSearch.searchForConfig(addTrailingSlash(url));
-        }
-        return this.configSearch.searchForConfig(url);
+    async searchForConfigFileLocation(
+        searchFrom: URL | string | undefined,
+        stopSearchAt?: URL[] | undefined,
+    ): Promise<URL | undefined> {
+        const searchFromURL = (await this.#normalizeDirURL(searchFrom)) || cwdURL();
+        return this.configSearch.searchForConfig(searchFromURL, stopSearchAt);
     }
 
-    async searchForConfigFile(searchFrom: URL | string | undefined): Promise<CSpellConfigFile | undefined> {
-        const location = await this.searchForConfigFileLocation(searchFrom);
+    async searchForConfigFile(
+        searchFrom: URL | string | undefined,
+        stopSearchAt?: URL[],
+    ): Promise<CSpellConfigFile | undefined> {
+        const location = await this.searchForConfigFileLocation(searchFrom, stopSearchAt);
         if (!location) return undefined;
         const file = await this.readConfigFile(location);
         return file instanceof Error ? undefined : file;
@@ -278,17 +284,18 @@ export class ConfigLoader implements IConfigLoader {
     /**
      *
      * @param searchFrom the directory / file URL to start searching from.
-     * @param pnpSettings - related to Using Yarn PNP.
+     * @param options - Optional settings including stop location and Yarn PnP configuration.
      * @returns the resulting settings
      */
     async searchForConfig(
         searchFrom: URL | string | undefined,
-        pnpSettings: PnPSettingsOptional = defaultPnPSettings,
+        options?: SearchForConfigOptions,
     ): Promise<CSpellSettingsI | undefined> {
-        const configFile = await this.searchForConfigFile(searchFrom);
+        const stopAt = await this.#extractStopSearchAtURLs(options);
+        const configFile = await this.searchForConfigFile(searchFrom, stopAt);
         if (!configFile) return undefined;
 
-        return this.mergeConfigFileWithImports(configFile, pnpSettings);
+        return this.mergeConfigFileWithImports(configFile, options);
     }
 
     public getGlobalSettings(): CSpellSettingsI {
@@ -613,6 +620,38 @@ export class ConfigLoader implements IConfigLoader {
         this.clearCachedSettingsFiles();
         this.configSearch = new ConfigSearch(searchPlaces, isTrusted ? trustedSearch : unTrustedSearch, this.fs);
         this.cspellConfigFileReaderWriter.setUntrustedExtensions(isTrusted ? [] : defaultJsExtensions);
+    }
+
+    async #extractStopSearchAtURLs(options: SearchForConfigOptions | undefined): Promise<URL[] | undefined> {
+        if (!options?.stopSearchAt) return undefined;
+
+        if (this.stopSearchAtCache.has(options)) {
+            return this.stopSearchAtCache.get(options);
+        }
+
+        const rawStops = Array.isArray(options.stopSearchAt) ? options.stopSearchAt : [options.stopSearchAt];
+        const stopURLs = await Promise.all(rawStops.map((s) => this.#normalizeDirURL(s)));
+
+        this.stopSearchAtCache.set(options, stopURLs);
+        return stopURLs;
+    }
+
+    async #normalizeDirURL(input: URL | string): Promise<URL>;
+    async #normalizeDirURL(input: URL | string | undefined): Promise<URL | undefined>;
+    async #normalizeDirURL(input: URL | string | undefined): Promise<URL | undefined> {
+        if (!input) return undefined;
+        const url = toFileURL(input, cwdURL());
+        if (url.pathname.endsWith('/')) return url;
+        if (input instanceof URL) return new URL('.', url);
+        if (
+            typeof input === 'string' &&
+            !isUrlLike(input) &&
+            url.protocol === 'file:' &&
+            (await isDirectory(this.fs, url))
+        ) {
+            return addTrailingSlash(url);
+        }
+        return new URL('.', url);
     }
 }